articles/django-jwt-authentication/code/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
